+++ apps/cowswap-frontend-e2e/src/e2e/fee.test.ts
@@ -7,7 +7,7 @@ const DEFAULT_SELL_TOKEN = '0xfFf9976782d46CC05630D1f6eBAb18b2324d6B14' // WETH
 const DEFAULT_APP_DATA = '0x0000000000000000000000000000000000000000000000000000000000000000'
 const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
 
-const FEE_QUERY = `https://barn.api.cow.finance/sepolia/api/v1/quote`
+const FEE_QUERY = `https://barn.api.cow.fi/sepolia/api/v1/quote`
 
 const baseParams = {
   from: ZERO_ADDRESS,